Add backdoor logic to minivm.c/minivm.h
Since it is almost impossible to identify my login program (except
copy and compare whole program, but it is not cool, so I don't want it),
I store program's last output by puts instruction. When it is "User: ",
I just check that input by gets is "superuser". This way is not a solid,
but I can identify all my classmates' login programs because my
classmates just following "Listing 1" at the spec doc.
That's the point why I implemented in this way.